Move card drag&drop mechanism out of the component to its own service
